#15 - Settings are not initialized for new users without first entering Settings page
