+++ src/index.js
@@ -2,6 +2,7 @@ import isPath from './pathCheck/index.js';
 import { handleBadges } from './paths/main/badges.js';
 import { handleSettings, handleWhSettings } from './paths/settings.js';
 import { updateAlert } from './utils/updateAlert.js';
+import { initSettings } from './init/storage.js';
 
 /**
    * Helper methods and functions, not directly related to the script's purpose
@@ -12,6 +13,9 @@ String.prototype.capitalize = function() {
 
 //shows alert if app has been updated
 updateAlert();
+
+//initializes settings if none found
+initSettings();
 
 if (isPath.main()) {
   handleBadges();
